requirements.txt
Flask>=2.0
openai>=1.0
pdfminer.six
python-docx
werkzeug